feat: S3 이미지일 경우에만 삭제처리하도록 업데이트
OAuth Server로부터 받은 profileImage의 경우 OAuth Server의 CDN에 저장되어 있는 이미지 파일로 완전한 url 형태이기 때문에 우리 S3에 저장하지 않았다.
따라서 S3에 존재하는 이미지일 경우에만 삭제 처리를 하도록 추가했다.

+++ backend/src/users/users.service.ts
@@ -108,7 +108,7 @@ export class UsersService {
 
         if (profileImage) {
             const curUserInfo = await this.findOne({ id: userId });
-            if (curUserInfo && curUserInfo.profileImage) {
+            if (curUserInfo && curUserInfo.profileImage && !curUserInfo.profileImage.startsWith('http')) {
                 await this.s3Service.deleteSingleObject(userId, curUserInfo.profileImage);
             }
         }
